Move Posts Collection frontend scripts to the collection package's frontend.js
